Support array from objects created outside the class

+++ runtime/memory/class_manager.cpp
@@ -47,7 +47,7 @@ std::pair<Field *, size_t> ClassManager::CreateFields(file_format::Class &asm_cl
         auto type = current_asm_field.GetType();
 
         PrintLog("relative_offset =  = ", relative_offset, "field_size = ", size,
-                 ", type = ", static_cast<int8_t>(type));
+                 ", type = ", static_cast<int16_t>(type));
 
         new (&runtime_fields[idx]) Field(type, size, relative_offset);
     }
